requirements.txt
streamlit
transformers
sentence-transformers
pymongo
pydantic
pillow
faiss-cpu
torch
torchvision
tritonclient[grpc]
numpy
pandas
requests
pytest
